// src/Model/oauthAccessTokenResponse.ts
/**
 * Pinterest REST API
 * Pinterest\'s REST API
 *
 * The version of the OpenAPI document: 5.13.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * A successful OAuth access token response.
 */
export interface OauthAccessTokenResponse {
    response_type?: OauthAccessTokenResponse.ResponseTypeEnum;
    access_token: string;
    token_type: string;
    expires_in: number;
    scope: string;
}

export namespace OauthAccessTokenResponse {
    export type ResponseTypeEnum = 'authorization_code' | 'refresh_token';
    export const ResponseTypeEnum = {
        AuthorizationCode: 'authorization_code' as ResponseTypeEnum,
        RefreshToken: 'refresh_token' as ResponseTypeEnum
    }
}