// src/Model/oauthAccessTokenResponseCode.ts
/**
 * Pinterest REST API
 * Pinterest\'s REST API
 *
 * The version of the OpenAPI document: 5.13.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OauthAccessTokenResponse } from './oauthAccessTokenResponse';


/**
 * A successful OAuth access token response for the authorization code flow.
 */
export interface OauthAccessTokenResponseCode extends OauthAccessTokenResponse {
    refresh_token: string;
    refresh_token_expires_in: number;
}

export namespace OauthAccessTokenResponseCode {

}